requirements.txt
absl-py==0.10.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==20.1.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==20.2.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==3.3.0
bmtk==0.0.8
bokeh==2.4.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
cloudpickle==2.0.0
colorednoise==1.1.1
cycler==0.10.0
dask==2022.3.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
distributed==2022.3.0
entrypoints==0.3
executing==0.8.2
flatbuffers==1.12
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gast==0.4.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
HeapDict==1.0.1
idna==2.10
imageio==2.9.0
ipykernel==5.3.4
ipympl==0.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.3
joblib==0.17.0
json5==0.9.5
jsons==1.4.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-packaging==0.10.1
jupyter-server==1.15.6
jupyterlab==3.2.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.1
jupyterlab-widgets==1.0.0
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libclang==12.0.0
llvmlite==0.38.0
locket==0.2.1
loss-landscapes==3.0.6
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.1.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.1.0
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.2.0
nest-asyncio==1.5.5
networkx==2.5
NEURON==7.8.2
notebook==6.4.10
numba==0.55.1
numpy==1.21.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.4.1
pandocfilters==1.4.3
parso==0.7.1
partd==1.2.0
pathspec==0.9.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==4.14.3
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.6.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.6.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==1.0.2
scipy==1.5.3
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.1.4
statsmodels==0.12.1
swifter==1.1.2
sympy==1.7.1
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.1.0
tifffile==2020.10.1
tomli==1.2.3
tomlkit==0.7.0
toolz==0.11.2
torch==1.10.0
torchmetrics==0.7.3
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing_extensions==4.0.1
typish==1.9.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.7.2
zict==2.1.0
